+++ deployment/keystone/changeset/contracts.go
@@ -194,43 +194,3 @@ func GetOwnedContractV2[T Ownable](addrs datastore.AddressRefStore, chain cldf_e
 
 	return ownedContract, nil
 }
-
-// loadCapabilityRegistry loads the CapabilitiesRegistry contract from the address book or datastore.
-func loadCapabilityRegistry(registryChain cldf_evm.Chain, env cldf.Environment, ref datastore.AddressRefKey) (*OwnedContract[*capabilities_registry.CapabilitiesRegistry], error) {
-	err := shouldUseDatastore(env, ref)
-	if err != nil {
-		return nil, fmt.Errorf("failed to check registry ref: %w", err)
-	}
-
-	var cr *OwnedContract[*capabilities_registry.CapabilitiesRegistry]
-
-	// `shouldUseDatastore` is already checking for the nil ref, no need to `ref == nil` here
-	a, err := env.DataStore.Addresses().Get(ref)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get address: %w", err)
-	}
-	cr, err = GetOwnedContractV2[*capabilities_registry.CapabilitiesRegistry](env.DataStore.Addresses(), registryChain, a.Address)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get owned contract: %w", err)
-	}
-
-	return cr, nil
-}
-
-func getTransferableContracts(addressStore datastore.AddressRefStore, chainSelector uint64) []common.Address {
-	var transferableContracts []common.Address
-
-	addresses := addressStore.Filter(datastore.AddressRefByChainSelector(chainSelector))
-	for _, addr := range addresses {
-		isOCR3Capability := addr.Type == datastore.ContractType(OCR3Capability)
-		isWorkflowRegistry := addr.Type == datastore.ContractType(WorkflowRegistry)
-		isKeystoneForwarder := addr.Type == datastore.ContractType(KeystoneForwarder)
-		isCapabilityRegistry := addr.Type == datastore.ContractType(CapabilitiesRegistry)
-
-		if isCapabilityRegistry || isWorkflowRegistry || isKeystoneForwarder || isOCR3Capability {
-			transferableContracts = append(transferableContracts, common.HexToAddress(addr.Address))
-		}
-	}
-
-	return transferableContracts
-}
